+++ src/universal/components/CreateCard/CreateCardRootStyles.js
@@ -10,7 +10,7 @@ const CreateCardRootStyles = {
   justifyContent: 'center',
   // TODO: Cards need block containers, not margin (TA)
   margin: '0 0 .5rem',
-  maxWidth: '20rem',
+  maxWidth: ui.cardMaxWidth,
   minHeight: ui.cardMinHeight,
   padding: '.6875rem 1.25rem .5rem'
 };