+++ src/universal/modules/userDashboard/containers/UserSettings/UserSettingsContainer.js
@@ -10,6 +10,7 @@ import {
 import {reduxForm, initialize} from 'redux-form';
 import {cashay} from 'cashay';
 import {withRouter} from 'react-router';
+import makeStep1Schema from 'universal/validation/makeStep1Schema';
 
 const updateSuccess = {
   title: 'Settings saved!',
@@ -25,9 +26,14 @@ const mapStateToProps = (state) => {
   };
 };
 
+const validate = (values) => {
+  const schema = makeStep1Schema();
+  return schema(values).errors;
+};
+
 @requireAuth
 @connect(mapStateToProps)
-@reduxForm({form: 'userSettings'})
+@reduxForm({form: 'userSettings', validate})
 @withRouter
 export default class UserSettingsContainer extends Component {
   static propTypes = {
@@ -41,6 +47,7 @@ export default class UserSettingsContainer extends Component {
       preferredName: PropTypes.string,
     }),
     router: PropTypes.object,
+    untouch: PropTypes.func.isRequired,
     userId: PropTypes.string
   };
 
@@ -49,7 +56,7 @@ export default class UserSettingsContainer extends Component {
   }
 
   onSubmit = (submissionData) => {
-    const {activity, dispatch, nextPage, user, userId, router} = this.props;
+    const {activity, dispatch, nextPage, untouch, user, userId, router} = this.props;
     const {preferredName} = submissionData;
     if (preferredName === user.preferredName) return;
     const options = {
@@ -68,6 +75,7 @@ export default class UserSettingsContainer extends Component {
     if (nextPage) {
       router.push(nextPage);
     }
+    untouch('preferredName');
   };
 
   initializeForm() {