+++ src/universal/routes/logout.js
@@ -1,7 +1,0 @@
-export default {
-  path: '/logout',
-  getComponent: async(location, cb) => {
-    const component = await System.import('universal/containers/Logout/LogoutContainer');
-    cb(null, component);
-  }
-};