+++ CHANGELOG.md
@@ -4,6 +4,23 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 This CHANGELOG follows conventions [outlined here](http://keepachangelog.com/).
 
+## v0.13.2 - 27-Nov-2016
+### Added
+- 'Player joined' message when teammate accepts invitation
+
+### Fixed
+- #543 unable to end action meeting
+
+## v0.13.1 - 25-Nov-2016
+### Added
+- Can now filter by team member on team dashboard Views
+- Updated FontAwesome to v4.7.0
+
+### Fixed
+- #514 Relabeling Actions and Projects during agenda processing
+  - Copy now reads "New private action" and "New team project"
+- #536 Cashay warning while proceeding through check-in round
+
 ## v0.13.0 - 22-Nov-2016
 We tagged v0.13.0 on our 1,300th commit. What a coinkidink!
 ### Added
@@ -69,7 +86,7 @@ We tagged v0.13.0 on our 1,300th commit. What a coinkidink!
 
 ### Fixed
 - #438 fixed TypeError: Cannot read property 'id' of undefined
-- #447 unable to logout and login properly
+- #447 unable to signout and login properly
 - Fixed race condition landing on meeting summary route at end of meeting
 - Fixed short urls